pyproject.toml
[project]
name = "project-weather-brazilian-capitals"
version = "0.1.0"
description = ""
authors = [
    {name = "Keityr",email = "seu@email.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
